feat: isolated the service layer creation logic
