+++ lib/src/utils/writeFile.js
@@ -1,4 +1,4 @@
-const fs = require('fs')
+const { writeFile } = require('fs')
 
 /**
  * @param {String} filename
@@ -7,7 +7,7 @@ const fs = require('fs')
  */
 module.exports = (filename, data) => {
   return new Promise((resolve, reject) => {
-    fs.writeFile(filename, data, error => {
+    writeFile(filename, data, error => {
       if (error) reject(error.message)
       else resolve('Saved successfully')
     })
